Split offscreen node opacity patching
